+++ src/RcppExports.cpp
@@ -2425,6 +2425,20 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// cpp_test_miid_prior
+List cpp_test_miid_prior(NumericVector theta_logchol, int p, int n, NumericVector x);
+RcppExport SEXP _tulpa_cpp_test_miid_prior(SEXP theta_logcholSEXP, SEXP pSEXP, SEXP nSEXP, SEXP xSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericVector >::type theta_logchol(theta_logcholSEXP);
+    Rcpp::traits::input_parameter< int >::type p(pSEXP);
+    Rcpp::traits::input_parameter< int >::type n(nSEXP);
+    Rcpp::traits::input_parameter< NumericVector >::type x(xSEXP);
+    rcpp_result_gen = Rcpp::wrap(cpp_test_miid_prior(theta_logchol, p, n, x));
+    return rcpp_result_gen;
+END_RCPP
+}
 // cpp_test_lkj_build_L
 List cpp_test_lkj_build_L(NumericVector raw, int n);
 RcppExport SEXP _tulpa_cpp_test_lkj_build_L(SEXP rawSEXP, SEXP nSEXP) {
@@ -3005,6 +3019,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_tulpa_cpp_test_s2z_block_schur", (DL_FUNC) &_tulpa_cpp_test_s2z_block_schur, 5},
     {"_tulpa_cpp_test_log_prior_icar", (DL_FUNC) &_tulpa_cpp_test_log_prior_icar, 7},
     {"_tulpa_cpp_test_mcar_prior", (DL_FUNC) &_tulpa_cpp_test_mcar_prior, 8},
+    {"_tulpa_cpp_test_miid_prior", (DL_FUNC) &_tulpa_cpp_test_miid_prior, 4},
     {"_tulpa_cpp_test_lkj_build_L", (DL_FUNC) &_tulpa_cpp_test_lkj_build_L, 2},
     {"_tulpa_cpp_test_lkj_density", (DL_FUNC) &_tulpa_cpp_test_lkj_density, 2},
     {"_tulpa_cpp_test_lkj_grad", (DL_FUNC) &_tulpa_cpp_test_lkj_grad, 3},